backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
agml==0.8.0
python-weather==1.1.1
python-dotenv==1.0.1
shap==0.44.0
scikit-learn==1.4.0
django-security==0.12.0
safety==2.3.5
bandit==1.7.5
pytest-django==4.5.2
pytest-cov==4.1.0